requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.1
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
pypdf==3.17.1
requests==2.31.0